+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.0] - 2022-26-11
+## Unreleased - 2023-22-02
+
+### Changed
+
+- Generic brute-force compute methods implement `ComputeMethod` directly using `Normed` trait.
+
+### Removed
+
+- Private `Computable` trait.
+
+## [0.4.0] - 2023-22-02
 
 ### Breaking changes